move personalization to its own file
